+++ pkg/armrpc/frontend/defaultoperation/getoperationresult.go
