+++ pkg/armrpc/frontend/defaultoperation/testdata/operationstatus_datamodel.json
