+++ pkg/armrpc/frontend/defaultoperation/testdata/operationstatus_output.json
